Add Dockerfile, update runPluginStarter.sh, and CI workflow
Added Dockerfile to create a containerized environment with pre-installed dependencies for running the plugin. This change will ensure ease of use and consistency across different host systems.

Updated runPluginStarter.sh to add default values and better error handling for arguments. This will improve the script's usability and robustness.

Updated GitHub Actions workflow (build.yml) to include building and pushing of Docker images to the GitHub Container Registry. This allows continuous delivery and deployment of changes made to the Dockerfile and the plugin.